pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
include = ["text_clustering/**/*.py", "paper/*.py"]

[project]
name = "text-clustering-llm"
version = "1.1.0"
description = "Text Clustering as Classification with LLMs — PPD M2 MLSD"
readme = "README.md"
requires-python = ">=3.12"
license = { text = "MIT" }
authors = [
  { name = "Akram Chaabnia" },
  { name = "Abdelheq Mokhtari" },
  { name = "Mohamed Raouf Razi" },
]

dependencies = [
  "openai>=1.30.0",
  "httpx>=0.27.0",
  "scikit-learn>=1.4.0",
  "scipy>=1.13.0",
  "numpy>=1.26.0",
  "tqdm>=4.66.0",
  "python-dotenv>=1.0.0",
]

[project.scripts]
tc-seed-labels    = "text_clustering.pipeline.seed_labels:main"
tc-label-gen      = "text_clustering.pipeline.label_generation:main_cli"
tc-classify       = "text_clustering.pipeline.classification:main_cli"
tc-evaluate       = "text_clustering.pipeline.evaluation:main_cli"

[project.optional-dependencies]
dev = [
  "commitizen>=3.27.0",
  "ruff>=0.4.0",
  "pre-commit>=3.7.0",
]

[tool.commitizen]
name = "cz_conventional_commits"
tag_format = "v$version"
version_scheme = "semver"
version_provider = "pep621"
update_changelog_on_bump = true
changelog_incremental = true
bump_message = "bump: release $current_version → $new_version [skip ci]"

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "W", "I"]
# Pipeline package: preserve original paper prompt strings (long lines, f-string style).
# Root shims are clean and need no per-file suppressions.
per-file-ignores = { "text_clustering/pipeline/label_generation.py" = ["E501", "F541"], "text_clustering/pipeline/classification.py" = ["E501", "F541"], "text_clustering/pipeline/evaluation.py" = ["E501", "E741"], "text_clustering/prompts.py" = ["E501", "F541"] }
